+++ app/common/src/services/RemoteBackend.ts
@@ -776,7 +776,7 @@ export class RemoteBackend extends backend.Backend {
           if (
             page.some((execution) => {
               const startDate = parseAbsolute(execution.startDate, execution.timeZone)
-              return startDate.year > year || startDate.month > month
+              return startDate.year > year && startDate.month > month
             })
           ) {
             break
@@ -1484,6 +1484,33 @@ export class RemoteBackend extends backend.Backend {
     return {
       projectRootId: backend.DirectoryId(`directory-${responseBody.projectRootDirectory}`),
       parentId: backend.DirectoryId(`directory-${responseBody.parentDirectory}`),
+    }
+  }
+
+  /** Downloads project session logs as a file. */
+  async downloadProjectSessionLogs(projectSessionId: backend.ProjectSessionId): Promise<void> {
+    const response = await this.get<{ readonly url: string }>(
+      remoteBackendPaths.getDownloadProjectSessionLogsPath(projectSessionId),
+    )
+
+    if (!response.ok) {
+      return await this.throw(response, 'getDownloadProjectSessionLogsBackendError')
+    } else {
+      const url = await response.json().then(
+        (json) => json.url,
+        () => null,
+      )
+      if (url == null) {
+        return await this.throw(response, 'getDownloadProjectSessionLogsBackendError')
+      }
+      await this.downloader({
+        url,
+        name: 'logs.txt',
+        electronOptions: {
+          path: null,
+        },
+      })
+      return
     }
   }